// web/src/types/carro.ts
export type Carro = {
    id: number;
    bodywork: string;
    carType: string;
    color: string;
    endPlate: number;
    fuelType: string;
    km: number;
    litersFuel: number;
    name: string;
    nameMarca: string;
    numberPassengers: number;
    price: number;
    qtDoors: number;
    trunkLiters: number;
    yearCar: number;
    bannerUrl: string;
    description: string;
}